httpredir: init
Add the configuration files for the HTTP to HTTPS redirector.
(i.e. redirect every HTTP request to the same request with HTTPS)
We do this with an extra container since HAProxy routes the path for
ACME challenges to a specific container and all other HTTP requests to
this container. It was problematic to let HAProxy also do the HTTPS
redirects because then the ACME challenges were also routed to this
container.